splash: Use larger icon again
https://gitlab.gnome.org/GNOME/adwaita-icon-theme/-/commit/a1c8a0f61e27e1bcda03412dec772b7a1ed5a795

removed the spinner icon, hence GTK3 added a one in

https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8354

but picks the 16x16 one which is a bit too small.

The 16x16 icon is fine for the other places we use a spinner, like
status pages.

This can go away with the switch to AdwSpinner.

Closes: https://gitlab.gnome.org/World/Phosh/phosh/-/issues/1222
Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1696>